refactor: GlobalErrorBoundary 삭제 후 일반 ErrorBoundary로 변경
